Fixed symlink creation logic

+++ switchVersion.sh
@@ -365,8 +365,9 @@ consoleWriteLine "Done."
 # the --fix-indexphp parameter.
 INDEX_PHP=$BASE/index.php
 INDEX_TARGET=$SYMLINK/index.php
+SYMLINK_TARGET=typo3_src/index.php
 consoleWriteVerbose "Checking if index.php needs to be updated..."
-if [[ -f "$INDEX_PHP" ]]; then
+if [[ -f "$INDEX_PHP" && ! -h "$INDEX_PHP" ]]; then
   rm -f "$INDEX_PHP"
   cp "$INDEX_TARGET" "$INDEX_PHP"
   consoleWriteLineVerbose "Done."
@@ -376,7 +377,7 @@ else
     consoleWriteLineVerbose "Skipped."
   else
     # If it doesn't exist, create a proper link.
-    ln -s "$INDEX_TARGET" "$INDEX_PHP"
+    ln -s "$SYMLINK_TARGET" "$INDEX_PHP"
     consoleWriteLineVerbose "Link created."
   fi
 fi
